+++ backend/src/main/java/org/example/backend/domain/fish/dto/FishRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.backend.domain.fish.dto.request;
+package org.example.backend.domain.fish.dto;
 
 public record FishRequestDto(
     String species,
